еще rest

+++ app/Http/Controllers/Admin/BrandsController.php
@@ -34,9 +34,10 @@ class BrandsController extends Controller
         return to_route('admin.brands.main');
     }
 
-    public function update(UpdateBrandRequest $request)
+    public function update(UpdateBrandRequest $request, Brand $brand)
     {
-        $brand = Brand::find($request->getData()['id']);
+        if($brand->id != $request->getData()['id']) abort(404);
+
         $brand->name = $request->getData()['name'];
         $brand->slug = $request->getData()['slug'];
         $brand->description = $request->getData()['description'];
@@ -59,9 +60,9 @@ class BrandsController extends Controller
         return to_route('admin.brands.main');
     }
 
-    public function delete(DeleteBrandRequest $request)
+    public function delete(DeleteBrandRequest $request, Brand $brand)
     {
-        $brand = Brand::find($request->getId());
+        if($brand->id != $request->getId()) abort(404);
 
         if(str_starts_with($brand->image_url, env('APP_URL', 'https://shop.slmatthew.ru').'/storage/')) {
             Storage::disk('public')->delete(mb_substr($brand->image_url, 9));
